+++ cmd/dockerd/daemon_systemd_unix.go
@@ -1,0 +1,23 @@
+//go:build linux && !no_systemd
+
+package main
+
+import (
+	"github.com/docker/docker/daemon"
+	"github.com/docker/docker/daemon/config"
+)
+
+func newCgroupParent(config *config.Config) string {
+	cgroupParent := "docker"
+	useSystemd := daemon.UsingSystemd(config)
+	if useSystemd {
+		cgroupParent = "system.slice"
+	}
+	if config.CgroupParent != "" {
+		cgroupParent = config.CgroupParent
+	}
+	if useSystemd {
+		cgroupParent = cgroupParent + ":" + "docker" + ":"
+	}
+	return cgroupParent
+}
